requirements.txt
streamlit>=1.28.0
pyyaml>=6.0
plotly>=5.17.0
